requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pypdfium2==5.4.0
python-dotenv==1.2.1
